+++ src/com/avona/games/towerdefence/TimeTrack.java
@@ -2,10 +2,10 @@ package com.avona.games.towerdefence;
 
 public class TimeTrack {
 
-	public double clock = 0;
-	public double tick = 0;
+	public float clock = 0;
+	public float tick = 0;
 
-	private double lastWallClock = 0;
+	private float lastWallClock = 0;
 	private boolean running = true;
 
 	public boolean isRunning() {
@@ -32,7 +32,7 @@ public class TimeTrack {
 	 * @param wallClock
 	 *            New wall clock time.
 	 */
-	public void update(double wallClock) {
+	public void update(float wallClock) {
 		if (lastWallClock != 0)
 			updateTick(wallClock - lastWallClock);
 		lastWallClock = wallClock;
@@ -47,21 +47,21 @@ public class TimeTrack {
 	 * @see TimeTrack#update(double)
 	 * @param wallTick
 	 */
-	public void updateTick(double wallTick) {
+	public void updateTick(float wallTick) {
 		if (running) {
 			tick = wallTick;
 			clock += wallTick;
 
-			if (wallTick < 0.00001) {
+			if (wallTick < 0.00001f) {
 				Util.log("zero tick..."); return;
 			}
 			// tickrate = (1.0 - wallTick) * tickrate + wallTick * (1.0 / wallTick);	
-			tickrate = (1.0 - wallTick) * tickrate + 1.0;
+			tickrate = (1.0f - wallTick) * tickrate + 1.0f;
 			// Util.log("tickrate: " + tickrate);
 		} else {
 			tick = 0;
 		}
 	}
 
-	public double tickrate = 60.0;
+	public float tickrate = 60.0f;
 }
